+++ tests/testthat-minor-transforms.R
@@ -1,4 +1,4 @@
 library(testthat)
 library(bestNormalize)
 
-test_check(package = 'bestNormalize', filter = "arcsinh|binarize|exp_x|log|no_transform|sqrt")
+test_check(package = 'bestNormalize', filter = "arcsinh|binarize|exp_x|log|no_transform|sqrt|center_scale")