+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Fixed error in documentation regarding `LambertW` transformation types 
   (thank you to Georg M. Goerg, the author of `LambertW`, for pointing this out). 
+- Add `center_scale` transform as default when `standardize == TRUE`
 
 # bestNormalize 1.6.1